refactor: separate Repositories into Interface and Implementation class
