+++ app/src/main/java/campus/tech/kakao/map/data/repository/SavedSearchWordRepository.kt
@@ -1,18 +1,11 @@
 package campus.tech.kakao.map.data.repository
 
-import campus.tech.kakao.map.data.SavedSearchWordDBHelper
 import campus.tech.kakao.map.model.SavedSearchWord
 
-class SavedSearchWordRepository(private val dbHelper: SavedSearchWordDBHelper) {
-    fun insertOrUpdateSearchWord(searchWord: SavedSearchWord) {
-        dbHelper.insertOrUpdateSearchWord(searchWord)
-    }
+interface SavedSearchWordRepository {
+    fun insertOrUpdateSearchWord(searchWord: SavedSearchWord)
 
-    fun getAllSearchWords(): List<SavedSearchWord> {
-        return dbHelper.getAllSearchWords()
-    }
+    fun getAllSearchWords(): List<SavedSearchWord>
 
-    fun deleteSearchWordById(id: Long) {
-        dbHelper.deleteSearchWordById(id)
-    }
+    fun deleteSearchWordById(id: Long)
 }
